rm a copy

+++ selfdrive/test/setup_device_ci.sh
@@ -80,7 +80,9 @@ safe_checkout() {
   echo "git checkout done, t=$SECONDS"
   du -hs $SOURCE_DIR $SOURCE_DIR/.git
 
-  rsync -a --delete $SOURCE_DIR $TEST_DIR
+  if [ ! -z "SKIP_COPY" ]; then
+    rsync -a --delete $SOURCE_DIR $TEST_DIR
+  fi
 }
 
 unsafe_checkout() {
